+++ app/renderer/components/layout/TopBar/TopBarContainer.js
@@ -5,11 +5,12 @@ import types from 'renderer/types';
 import { compose, lifecycle, withState } from 'recompose';
 import TopBar from ".";
 
-const mapStateToProps = ({ nodeSystem, remoteStream, syncStream, syncRemoteStream }) => ({
+const mapStateToProps = ({ nodeSystem, remoteStream, syncStream, syncRemoteStream, networkStatusStore }) => ({
   nodeSystem,
   remoteStream,
   syncStream,
   syncRemoteStream,
+  networkStatusStore,
 });
 
 const mapDispatchToProps = dispatch => ({