+++ app/renderer/pages/syncNodePage/container.js
@@ -31,7 +31,7 @@ const mapDispatchToProps = dispatch => ({
 
 const enhance = lifecycle({
   componentDidMount() {
-    const { localStorage, onRestartNode } = this.props;
+    const { localStorage, onRestartNode, nodeSystem } = this.props;
     const selectedNetwork = localStorage[storageKeys.SELECTED_NETWORK];
     const genesisConfigFile = localStorage[storageKeys.GENESIS_CONFIG_FILE_INFO];
     const genesisConfigFilePath = genesisConfigFile && genesisConfigFile.path;
@@ -40,7 +40,13 @@ const enhance = lifecycle({
         selectedNetwork.value === NetworkNameOptions.LOCAL_TESTNET && genesisConfigFilePath
           ? genesisConfigFilePath
           : selectedNetwork.value;
-      onRestartNode({ chain: targetChain });
+      const currentNetwork = NetworkNameMapping[nodeSystem.localNode.chain];
+      if(currentNetwork !== targetChain) {
+        Logger.debug(`restart node to use ${targetChain}`);
+        onRestartNode({ chain: targetChain });
+      } else {
+        Logger.debug(`same network skip restart. ${currentNetwork}`);
+      }
     } else {
       onRestartNode();
     }